Resolving comments. clearing task events head
