+++ python/ray/dashboard/modules/task_events/tests/test_task_events_head.py
@@ -1,4 +1,3 @@
-import asyncio
 import sys
 from unittest.mock import AsyncMock
 
@@ -10,14 +9,13 @@ from ray._common.ray_constants import (
     LOGGING_ROTATE_BACKUP_COUNT,
     LOGGING_ROTATE_BYTES,
 )
-from ray._common.test_utils import async_wait_for_condition
 from ray._raylet import JobID
 from ray.core.generated import (
     events_event_aggregator_service_pb2,
     gcs_pb2,
     gcs_service_pb2,
 )
-from ray.core.generated.common_pb2 import TaskStatus, TaskType
+from ray.core.generated.common_pb2 import TaskType
 from ray.core.generated.events_base_event_pb2 import RayEvent
 from ray.core.generated.events_task_definition_event_pb2 import TaskDefinitionEvent
 from ray.dashboard.modules.task_events import task_event_query
@@ -77,6 +75,14 @@ def _fake_request(body: bytes):
     request = AsyncMock()
     request.read.return_value = body
     return request
+
+
+def _add_stored_task(head, task_id, worker=None, job=_JOB):
+    event = gcs_pb2.TaskEvents(task_id=task_id, attempt_number=0, job_id=job)
+    event.task_info.type = TaskType.NORMAL_TASK
+    if worker is not None:
+        event.state_updates.worker_id = worker
+    head._store.add_or_replace_task_event(event)
 
 
 @pytest.mark.asyncio
@@ -165,185 +171,6 @@ async def test_get_task_events_invalid_predicate_returns_status_in_reply():
     reply.ParseFromString(response.body)
     assert reply.status.code == task_event_query._INVALID_ARGUMENT_STATUS_CODE
     assert len(reply.events_by_task) == 0
-
-
-_HEAD = "ray.dashboard.modules.task_events.task_events_head"
-_WORKER = b"worker_1"
-
-
-def _add_stored_task(head, task_id, worker=None, job=_JOB):
-    event = gcs_pb2.TaskEvents(task_id=task_id, attempt_number=0, job_id=job)
-    event.task_info.type = TaskType.NORMAL_TASK
-    if worker is not None:
-        event.state_updates.worker_id = worker
-    head._store.add_or_replace_task_event(event)
-
-
-def _is_failed(head, task_id) -> bool:
-    task_event = head._store.get_task_event((task_id, 0))
-    return (
-        task_event is not None
-        and TaskStatus.FAILED in task_event.state_updates.state_ts_ns
-    )
-
-
-@pytest.mark.asyncio
-async def test_handle_worker_delta_fails_tasks(monkeypatch):
-    monkeypatch.setattr(f"{_HEAD}._MARK_FAILED_ON_WORKER_DEAD_DELAY_S", 0.0)
-    head = _make_head()
-    _add_stored_task(head, _task_id(1), worker=_WORKER)
-
-    async def fake_get(worker_id):
-        assert worker_id == _WORKER
-        return gcs_pb2.WorkerTableData(exit_detail="boom", end_time_ms=5)
-
-    monkeypatch.setattr(head, "_get_worker_info", fake_get)
-
-    head._handle_worker_delta(gcs_pb2.WorkerDeltaData(worker_id=_WORKER))
-
-    await async_wait_for_condition(lambda: _is_failed(head, _task_id(1)))
-
-
-@pytest.mark.asyncio
-async def test_handle_worker_delta_missing_worker_info_is_noop(monkeypatch):
-    monkeypatch.setattr(f"{_HEAD}._MARK_FAILED_ON_WORKER_DEAD_DELAY_S", 0.0)
-    head = _make_head()
-    _add_stored_task(head, _task_id(1), worker=_WORKER)
-
-    async def fake_get(worker_id):
-        return None
-
-    monkeypatch.setattr(head, "_get_worker_info", fake_get)
-
-    head._handle_worker_delta(gcs_pb2.WorkerDeltaData(worker_id=_WORKER))
-    await asyncio.sleep(0.05)
-
-    assert not _is_failed(head, _task_id(1))
-
-
-@pytest.mark.asyncio
-async def test_handle_job_update_finished_fails_tasks(monkeypatch):
-    monkeypatch.setattr(f"{_HEAD}._MARK_FAILED_ON_JOB_DONE_DELAY_S", 0.0)
-    head = _make_head()
-    _add_stored_task(head, _task_id(1))
-
-    head._handle_job_update(gcs_pb2.JobTableData(job_id=_JOB, is_dead=True, end_time=5))
-
-    await async_wait_for_condition(lambda: _is_failed(head, _task_id(1)))
-
-
-@pytest.mark.asyncio
-async def test_handle_job_update_ignores_running_job():
-    head = _make_head()
-    _add_stored_task(head, _task_id(1))
-
-    head._handle_job_update(gcs_pb2.JobTableData(job_id=_JOB, is_dead=False))
-    await asyncio.sleep(0.05)
-
-    assert not _is_failed(head, _task_id(1))
-
-
-class _FakeSubscriber:
-    """Delivers a canned batch on the first poll, then parks so the loop settles."""
-
-    def __init__(self, messages):
-        self._messages = messages
-        self._delivered = False
-
-    async def subscribe(self):
-        pass
-
-    async def poll(self, batch_size, timeout=None):
-        if self._delivered:
-            await asyncio.sleep(3600)
-            return []
-        self._delivered = True
-        return self._messages
-
-
-async def _cancel(task: asyncio.Task) -> None:
-    task.cancel()
-    try:
-        await task
-    except asyncio.CancelledError:
-        pass
-
-
-@pytest.mark.asyncio
-async def test_gc_job_summary_loop_trims_over_cap(monkeypatch):
-    monkeypatch.setattr(
-        "ray.dashboard.modules.task_events.task_events_head.GC_JOB_SUMMARY_INTERVAL_S",
-        0.01,
-    )
-    monkeypatch.setattr(
-        "ray.dashboard.modules.task_events.task_event_storage."
-        "MAX_DROPPED_TASK_ATTEMPTS_PER_JOB",
-        2,
-    )
-    head = _make_head()
-    summary = head._store._summary(_JOB)
-    for i in range(10):
-        summary.record_task_attempt_dropped((_task_id(i), 0))
-    assert len(summary._dropped_task_attempts) == 10
-
-    task = asyncio.create_task(head._gc_job_summary_loop())
-    try:
-        await async_wait_for_condition(lambda: len(summary._dropped_task_attempts) < 10)
-    finally:
-        await _cancel(task)
-
-
-@pytest.mark.asyncio
-async def test_worker_death_subscription_loop_reconciles(monkeypatch):
-    monkeypatch.setattr(f"{_HEAD}._MARK_FAILED_ON_WORKER_DEAD_DELAY_S", 0.0)
-    head = _make_head()
-    _add_stored_task(head, _task_id(1), worker=_WORKER)
-
-    async def fake_get(worker_id):
-        return gcs_pb2.WorkerTableData(exit_detail="boom", end_time_ms=5)
-
-    monkeypatch.setattr(head, "_get_worker_info", fake_get)
-
-    fake = _FakeSubscriber([(b"key", gcs_pb2.WorkerDeltaData(worker_id=_WORKER))])
-    monkeypatch.setattr(
-        f"{_HEAD}.GcsAioWorkerDeltaSubscriber", lambda address=None: fake
-    )
-
-    task = asyncio.create_task(head._subscribe_for_worker_deaths())
-    try:
-        await async_wait_for_condition(lambda: _is_failed(head, _task_id(1)))
-    finally:
-        await _cancel(task)
-
-
-@pytest.mark.asyncio
-async def test_job_subscription_loop_reconciles_only_finished(monkeypatch):
-    monkeypatch.setattr(f"{_HEAD}._MARK_FAILED_ON_JOB_DONE_DELAY_S", 0.0)
-    head = _make_head()
-    running_job = JobID.from_int(1).binary()
-    finished_job = JobID.from_int(2).binary()
-    _add_stored_task(head, _task_id(1), job=running_job)
-    _add_stored_task(head, _task_id(2), job=finished_job)
-
-    fake = _FakeSubscriber(
-        [
-            (b"k1", gcs_pb2.JobTableData(job_id=running_job, is_dead=False)),
-            (
-                b"k2",
-                gcs_pb2.JobTableData(job_id=finished_job, is_dead=True, end_time=5),
-            ),
-        ]
-    )
-    monkeypatch.setattr(f"{_HEAD}.GcsAioJobSubscriber", lambda address=None: fake)
-
-    task = asyncio.create_task(head._subscribe_for_finished_jobs())
-    try:
-        await async_wait_for_condition(lambda: _is_failed(head, _task_id(2)))
-    finally:
-        await _cancel(task)
-
-    # The running job was ignored; its task is not failed.
-    assert not _is_failed(head, _task_id(1))
 
 
 if __name__ == "__main__":
